+++ adwords_downloader/config.py
@@ -42,3 +42,15 @@ def oauth2_refresh_token() -> str:
 def redownload_window() -> str:
     """The number of days for which the performance data will be redownloaded"""
     return '30'
+
+def output_file_version() -> str:
+    """A suffix that is added to output files, denoting a version of the data format"""
+    return 'v3'
+
+def max_retries() -> int:
+    """How often try retry at max in case of 500 errors"""
+    return 5
+
+def retry_backoff_factor() -> int:
+    """How many seconds to wait between retries (is multiplied with retry count)"""
+    return 5