+++ adwords_downloader/downloader.py
@@ -17,7 +17,7 @@ from googleads import adwords, oauth2, errors
 from oauth2client import client as oauth2_client
 
 API_VERSION = 'v201705'
-OUTPUT_FILE_VERSION = 'v1'
+
 
 # Timeout between retries in seconds.
 BACKOFF_FACTOR = 5
@@ -144,7 +144,7 @@ def download_performance(api_client: AdWordsApiClient,
         relative_filepath = Path('{date:%Y/%m/%d}/adwords/{filename}_{version}.json.gz'.format(
             date=current_date,
             filename=performance_report_type.value,
-            version=OUTPUT_FILE_VERSION))
+            version=config.output_file_version()))
         filepath = ensure_data_directory(relative_filepath)
 
         if (not filepath.is_file()
@@ -206,7 +206,7 @@ def download_account_structure(api_client: AdWordsApiClient):
         api_client: An AdWordsApiClient
 
     """
-    filename = Path('adwords-account-structure_{}.csv.gz'.format(OUTPUT_FILE_VERSION))
+    filename = Path('adwords-account-structure_{}.csv.gz'.format(config.output_file_version()))
     filepath = ensure_data_directory(filename)
 
     with tempfile.TemporaryDirectory() as tmp_dir:
@@ -398,12 +398,13 @@ def _download_adwords_report(api_client: AdWordsApiClient,
                                                               skip_report_summary=False)
             return report
         except errors.AdWordsReportError as e:
-            if e.code == 500 and retry_count < MAX_RETRIES:
+            if e.code == 500 and retry_count < config.max_retries():
                 logging.warning(('Failed attempt #{retry_count} for report with settings:\n'
                                  '{report_filter}\n'
                                  'Retrying...').format(retry_count=retry_count,
                                                        report_filter=report_filter))
-                time.sleep(retry_count * BACKOFF_FACTOR)
+
+                time.sleep(retry_count * config.retry_backoff_factor())
             else:
                 raise e