+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+
+
+
+## 1.2.0 
+*2017-09-20* 
+
+- Updated googleads-python-lib to 6.0 and use AdWords API version v201705
+- Added retry logic
+- Made the config and click commands discoverable in [mara-app](https://github.com/mara/mara-app) >= 1.2.0
+
+**required changes**
+
+- The file format changed to `v3`. Adapt etl scripts that process the output data.
+
+
 ## 1.1.1
 *2017-06-30* 
 
@@ -14,3 +29,4 @@
 *2017-02-24* 
 
 - Initial version
+